refactor: restructure constants.ts to Constants.ts
Split the monolithic constants.ts into a structured Constants.ts aggregator with separate modules for internal IDs and user-facing messages.

1. Structure Changes:
   - Rename src/constants.ts to src/Constants.ts (PascalCase).
   - Extract internal constants to src/constants/Ids.ts.
   - Extract user-facing strings to src/constants/Messages.ts.

2. Refactoring:
   - Update all imports across the codebase to use the new Constants module.
   - Fix dynamic import in RegexUtils.ts to resolve the new path.

This improves maintainability by separating concerns and following the project's file naming conventions.
