+++ src/services/ColorService.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { Constants } from '../constants';
+import { Constants } from '../Constants';
 import { FilterGroup } from '../models/Filter';
 
 // Solarized-inspired and distinct colors for highlights
